refactor out the Manifest and Lockfile classes
